# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Beratungssuche REST-API",
    "description": "Backend der Beratungssuche TU Berlin.\n",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "matildafreiheit@tu-berlin.de"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://virtserver.swaggerhub.com/bachelorarbeit-c90/Bachelorarbeit/1.0.0",
      "description": "SwaggerHub API Auto Mocking"
    },
    {
      "url": "https://vm021.qu.tu-berlin.de:3000",
      "description": "VM TU Berlin"
    },
    {
      "url": "https://vm021.qu.tu-berlin.de:3000"
    }
  ],
...
  "paths": {
    "/angebote": {
      "get": {
        "tags": [
          "angebote"
        ],
        "summary": "Finde alle Angebote",
        "operationId": "getAllAngebote",
        "parameters": [
          {
            "name": "lang",
            "in": "query",
            "description": "Sprache der Ergebnisse (de = Deutsch, en = Englisch).",
            "required": false,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string",
              "enum": [
                "de",
                "en"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Alle Angebote erfolgreich abgerufen",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Angebot"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Fehler beim Abrufen der Angebote."
          }
        }
      },
      "post": {
        "tags": [
          "angebote"
        ],
        "summary": "Neues Angebot erstellen",
        "description": "Fügt ein neues Angebot hinzu.",
        "operationId": "postAngebot",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Angebot"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Angebot erfolgreich erstellt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_201"
                }
              }
            }
          },
          "400": {
            "description": "Fehlende Felder oder ungültige Anfrage"
          },
          "500": {
            "description": "Fehler beim Erstellen des Angebots."
          }
        }
      }
    },
    "/angebote/{angebotId}": {
      "get": {
        "tags": [
          "angebote"
        ],
        "summary": "Suche Angebot nach ID",
        "description": "liefert ein Angebot zurueck",
        "operationId": "getOneAngebot",
        "parameters": [
          {
            "name": "angebotId",
            "in": "path",
            "description": "ID des Angebotes, nach dem gesucht wird",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "lang",
            "in": "query",
            "description": "Sprache der Ergebnisse (de = Deutsch, en = Englisch).",
            "required": false,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string",
              "enum": [
                "de",
                "en"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Angebot gefunden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Angebot"
                }
              }
            }
          },
          "404": {
            "description": "Angebot nicht gefunden"
          },
          "500": {
            "description": "Fehler beim Abrufen des Angebots."
          }
        }
      },
      "put": {
        "tags": [
          "angebote"
        ],
        "summary": "Angebot aktualisieren",
        "description": "Aktualisiert die Details eines bestehenden Angebots.",
        "operationId": "updateAngebot",
        "parameters": [
          {
            "name": "angebotId",
            "in": "path",
            "description": "ID des Angebots",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Angebot"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Angebot erfolgreich aktualisiert"
          },
          "400": {
            "description": "Fehlende Felder oder ungültige Anfrage"
          },
          "404": {
            "description": "Angebot nicht gefunden"
          },
          "500": {
            "description": "Fehler beim Aktualisieren des Angebots."
          }
        }
      },
      "delete": {
        "tags": [
          "angebote"
        ],
        "summary": "Angebot löschen",
        "description": "Löscht ein Angebot basierend auf seiner ID.",
        "operationId": "deleteAngebot",
        "parameters": [
          {
            "name": "angebotId",
            "in": "path",
            "description": "ID des Angebots",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Angebot erfolgreich gelöscht",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_200"
                }
              }
            }
          },
          "404": {
            "description": "Angebot nicht gefunden"
          },
          "500": {
            "description": "Fehler beim Löschen des Angebots."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Angebot": {
        "type": "object",
        "properties": {
          "ID": {
            "type": "integer"
          },
          "InstitutionID": {
            "type": "integer"
          },
          "Name": {
            "type": "string"
          },
          "Beschreibung": {
            "type": "string"
          },
          "url": {
            "type": "string",
            "format": "uri"
          },
          "TagIDs": {
            "type": "array",
            "items": {
              "type": "integer"
            }
          },
          "ZielgruppenIDs": {
            "type": "array",
            "items": {
              "type": "integer"
            }
          },
          "Arten": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "Online",
                "Campus",
                "Website",
                "Telefon"
              ]
            }
          }
        },
        "xml": {
          "name": "Angebot"
        }
      },
...
      "inline_response_201": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "institutionId": {
            "type": "integer"
          },
          "angebotId": {
            "type": "integer"
          }
        }
      },
      "inline_response_200": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        },
        "example": {
          "success": true,
          "message": "Angebot erfolgreich gelöscht."
        }
      },
...
    },
...
  }
}